+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Update go version from 1.18.1 to 1.18.6
+- Update Makefiles to version 7.0.1
+- Update archive package to handle archives in memory (#16)
+
+### Fixed
+- Fix date of files added to archive (#14)
 
 ## [v0.5.0] - 2022-09-12
 ### Added